third_party/WebKit/LayoutTests/editing/selection/mouse/long-press-on-hidden-node-crash.html

<!doctype HTML>
<html hidden id="html">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
test(function () {
    assert_not_equals(window.eventSender, undefined, 'This test requires eventSender.'); 
    var x = html.offsetLeft + 10;
    var y = html.offsetTop + 40;
    eventSender.gestureLongPress(x, y);

    html.hidden = "";
});
</script>
</html>
